feat: add basic pairing wizard UI
